pyproject.toml
[tool.poetry]
name = "pytkquiz"
version = "0.1.0"
description = ""
authors = ["Richard Anton <ranton@ranton.org>"]
readme = "README.md"

# had to do pip install playsound --only-binary playsound for 3.12 on mac
[tool.poetry.dependencies]
python = "^3.11.7"
pillow = "*"
gtts = "*"
# playsound = "*"
PyObjC = "*"
wheel = "0.43.0"
types-playsound = "^1.3.1.3"
streamlit = "^1.4.0"

[tool.poetry-exec-plugin.commands]
hackydeps = "pip install playsound --only-binary playsound"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
flake8 = "^7.0.0"
flake8-unused-arguments = "^0.0.13"
isort = "^5.13.2"
black = "^23.12.1"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
